Add new symbol and footprint for DC2DC converter
In order to power PT which requires 8V ~ 12V power supply, an 5V_12V
DC2DC converter in added into the schematic. Since the converter is
custom manufactured by an independent company, the correspond data
is not included in the Kicad library. So I have to draw it by my self.

Now, both symbol and footprint are finished, the next step is to
redraw the whole PCB layout and make it even more compact.